+++ adapters/blasto/blastotest/exemplary/native-app.json
@@ -27,14 +27,15 @@
       {
         "id": "some-impression-id",
         "tagid": "ogTAGID",
-        "banner": {
-          "w": 320,
-          "h": 50
+        "native": {
+          "ver": "1.1",
+          "request": "{\"adunit\":2,\"assets\":[{\"id\":3,\"img\":{\"h\":120,\"hmin\":0,\"type\":3,\"w\":180,\"wmin\":0},\"required\":1},{\"id\":0,\"required\":1,\"title\":{\"len\":25}},{\"data\":{\"len\":25,\"type\":1},\"id\":4,\"required\":1},{\"data\":{\"len\":140,\"type\":2},\"id\":6,\"required\":1}],\"context\":1,\"layout\":1,\"contextsubtype\":11,\"plcmtcnt\":1,\"plcmttype\":2,\"ver\":\"1.1\",\"ext\":{\"banner\":{\"w\":320,\"h\":50}}}"
         },
         "ext": {
           "bidder": {
             "accountId": "accountId",
-            "placementId": "placementId"
+            "sourceId": "sourceId",
+            "host": "host"
           }
         }
       }
@@ -60,7 +61,7 @@
             "123.123.123.123"
           ]
         },
-        "uri": "http://us-e-node1.bizzclick.com/bid?rtb_seat_id=placementId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "device": {
@@ -72,9 +73,9 @@
           "imp": [
             {
               "id": "some-impression-id",
-              "banner": {
-                "w": 320,
-                "h": 50
+              "native": {
+                "ver": "1.1",
+                "request": "{\"adunit\":2,\"assets\":[{\"id\":3,\"img\":{\"h\":120,\"hmin\":0,\"type\":3,\"w\":180,\"wmin\":0},\"required\":1},{\"id\":0,\"required\":1,\"title\":{\"len\":25}},{\"data\":{\"len\":25,\"type\":1},\"id\":4,\"required\":1},{\"data\":{\"len\":140,\"type\":2},\"id\":6,\"required\":1}],\"context\":1,\"layout\":1,\"contextsubtype\":11,\"plcmtcnt\":1,\"plcmttype\":2,\"ver\":\"1.1\",\"ext\":{\"banner\":{\"w\":320,\"h\":50}}}"
               },
               "tagid": "ogTAGID"
             }
@@ -96,7 +97,9 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
         "status": 200,
@@ -107,19 +110,17 @@
               "bid": [
                 {
                   "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-                  "impid": "1",
+                  "impid": "some-impression-id",
                   "price": 3.5,
                   "adm": "awesome-markup",
                   "adomain": [
                     "awesome.com"
                   ],
-                  "crid": "20",
-                  "w": 320,
-                  "h": 50
+                  "crid": "20"
                 }
               ],
-              "type": "banner",
-              "seat": "bizzclick"
+              "type": "native",
+              "seat": "blasto"
             }
           ],
           "cur": "USD"
@@ -133,17 +134,15 @@
         {
           "bid": {
             "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-            "impid": "1",
+            "impid": "some-impression-id",
             "price": 3.5,
             "adm": "awesome-markup",
+            "crid": "20",
             "adomain": [
               "awesome.com"
-            ],
-            "crid": "20",
-            "w": 320,
-            "h": 50
+            ]
           },
-          "type": "banner"
+          "type": "native"
         }
       ]
     }
