+++ adapters/blasto/blastotest/exemplary/video-app.json
@@ -11,34 +11,30 @@
     "user": {
       "buyeruid": "awesome-user"
     },
-    "site": {
-      "page": "test.com",
+    "app": {
       "publisher": {
         "id": "123456789"
-      }
+      },
+      "cat": [
+        "IAB22-1"
+      ],
+      "bundle": "com.app.awesome",
+      "name": "Awesome App",
+      "domain": "awesomeapp.com",
+      "id": "123456789"
     },
     "imp": [
       {
-        "id": "some-impression-id1",
+        "id": "some-impression-id",
         "tagid": "ogTAGID",
-        "banner": {
-          "w": 320,
-          "h": 50
-        },
-        "ext": {
-          "bidder": {
-            "accountId": "accountId",
-            "sourceId": "sourceId",
-            "host": "host"
-          }
-        }
-      },
-      {
-        "id": "some-impression-id2",
-        "tagid": "ogTAGID",
-        "banner": {
-          "w": 320,
-          "h": 50
+        "video": {
+          "mimes": [
+            "video/mp4"
+          ],
+          "w": 640,
+          "h": 480,
+          "minduration": 120,
+          "maxduration": 150
         },
         "ext": {
           "bidder": {
@@ -70,7 +66,7 @@
             "123.123.123.123"
           ]
         },
-        "uri": "http://host.bizzclick.com/bid?rtb_seat_id=sourceId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "device": {
@@ -81,24 +77,27 @@
           },
           "imp": [
             {
-              "id": "some-impression-id1",
-              "tagid": "ogTAGID",
-              "banner": {
-                "w": 320,
-                "h": 50
-              }
-            },
-            {
-              "id": "some-impression-id2",
-              "tagid": "ogTAGID",
-              "banner": {
-                "w": 320,
-                "h": 50
-              }
+              "id": "some-impression-id",
+              "video": {
+                "mimes": [
+                  "video/mp4"
+                ],
+                "minduration": 120,
+                "maxduration": 150,
+                "w": 640,
+                "h": 480
+              },
+              "tagid": "ogTAGID"
             }
           ],
-          "site": {
-            "page": "test.com",
+          "app": {
+            "id": "123456789",
+            "name": "Awesome App",
+            "bundle": "com.app.awesome",
+            "domain": "awesomeapp.com",
+            "cat": [
+              "IAB22-1"
+            ],
             "publisher": {
               "id": "123456789"
             }
@@ -108,7 +107,9 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id1","some-impression-id2"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
         "status": 200,
@@ -119,31 +120,18 @@
               "bid": [
                 {
                   "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-                  "impid": "some-impression-id1",
+                  "impid": "some-impression-id",
                   "price": 3.5,
                   "adm": "awesome-markup",
                   "adomain": [
                     "awesome.com"
                   ],
                   "crid": "20",
-                  "w": 320,
-                  "h": 50
-                },
-                {
-                  "id": "a3ae1b4e2fc24a4fb45540082e98e162",
-                  "impid": "some-impression-id2",
-                  "price": 3.5,
-                  "adm": "awesome-markup",
-                  "adomain": [
-                    "awesome.com"
-                  ],
-                  "crid": "20",
-                  "w": 320,
-                  "h": 50
+                  "w": 1280,
+                  "h": 720
                 }
               ],
-              "type": "banner",
-              "seat": "bizzclick"
+              "seat": "blasto"
             }
           ],
           "cur": "USD"
@@ -157,32 +145,17 @@
         {
           "bid": {
             "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-            "impid": "some-impression-id1",
+            "impid": "some-impression-id",
             "price": 3.5,
             "adm": "awesome-markup",
             "crid": "20",
             "adomain": [
               "awesome.com"
             ],
-            "w": 320,
-            "h": 50
+            "w": 1280,
+            "h": 720
           },
-          "type": "banner"
-        },
-        {
-          "bid": {
-            "id": "a3ae1b4e2fc24a4fb45540082e98e162",
-            "impid": "some-impression-id2",
-            "price": 3.5,
-            "adm": "awesome-markup",
-            "crid": "20",
-            "adomain": [
-              "awesome.com"
-            ],
-            "w": 320,
-            "h": 50
-          },
-          "type": "banner"
+          "type": "video"
         }
       ]
     }
