+++ adapters/blasto/blastotest/exemplary/video-web.json
@@ -11,25 +11,24 @@
     "user": {
       "buyeruid": "awesome-user"
     },
-    "app": {
+    "site": {
+      "page": "test.com",
       "publisher": {
         "id": "123456789"
-      },
-      "cat": [
-        "IAB22-1"
-      ],
-      "bundle": "com.app.awesome",
-      "name": "Awesome App",
-      "domain": "awesomeapp.com",
-      "id": "123456789"
+      }
     },
     "imp": [
       {
         "id": "some-impression-id",
         "tagid": "ogTAGID",
-        "native": {
-          "ver": "1.1",
-          "request": "{\"adunit\":2,\"assets\":[{\"id\":3,\"img\":{\"h\":120,\"hmin\":0,\"type\":3,\"w\":180,\"wmin\":0},\"required\":1},{\"id\":0,\"required\":1,\"title\":{\"len\":25}},{\"data\":{\"len\":25,\"type\":1},\"id\":4,\"required\":1},{\"data\":{\"len\":140,\"type\":2},\"id\":6,\"required\":1}],\"context\":1,\"layout\":1,\"contextsubtype\":11,\"plcmtcnt\":1,\"plcmttype\":2,\"ver\":\"1.1\",\"ext\":{\"banner\":{\"w\":320,\"h\":50}}}"
+        "video": {
+          "mimes": [
+            "video/mp4"
+          ],
+          "w": 640,
+          "h": 480,
+          "minduration": 120,
+          "maxduration": 150
         },
         "ext": {
           "bidder": {
@@ -61,7 +60,7 @@
             "123.123.123.123"
           ]
         },
-        "uri": "http://host.bizzclick.com/bid?rtb_seat_id=sourceId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "device": {
@@ -73,21 +72,20 @@
           "imp": [
             {
               "id": "some-impression-id",
-              "native": {
-                "ver": "1.1",
-                "request": "{\"adunit\":2,\"assets\":[{\"id\":3,\"img\":{\"h\":120,\"hmin\":0,\"type\":3,\"w\":180,\"wmin\":0},\"required\":1},{\"id\":0,\"required\":1,\"title\":{\"len\":25}},{\"data\":{\"len\":25,\"type\":1},\"id\":4,\"required\":1},{\"data\":{\"len\":140,\"type\":2},\"id\":6,\"required\":1}],\"context\":1,\"layout\":1,\"contextsubtype\":11,\"plcmtcnt\":1,\"plcmttype\":2,\"ver\":\"1.1\",\"ext\":{\"banner\":{\"w\":320,\"h\":50}}}"
-              },
-              "tagid": "ogTAGID"
+              "tagid": "ogTAGID",
+              "video": {
+                "mimes": [
+                  "video/mp4"
+                ],
+                "minduration": 120,
+                "maxduration": 150,
+                "w": 640,
+                "h": 480
+              }
             }
           ],
-          "app": {
-            "id": "123456789",
-            "name": "Awesome App",
-            "bundle": "com.app.awesome",
-            "domain": "awesomeapp.com",
-            "cat": [
-              "IAB22-1"
-            ],
+          "site": {
+            "page": "test.com",
             "publisher": {
               "id": "123456789"
             }
@@ -97,7 +95,9 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
         "status": 200,
@@ -114,11 +114,17 @@
                   "adomain": [
                     "awesome.com"
                   ],
-                  "crid": "20"
+                  "crid": "20",
+                  "w": 1280,
+                  "h": 720,
+                  "ext": {
+                    "prebid": {
+                      "type": "video"
+                    }
+                  }
                 }
               ],
-              "type": "native",
-              "seat": "bizzclick"
+              "seat": "blasto"
             }
           ],
           "cur": "USD"
@@ -135,12 +141,19 @@
             "impid": "some-impression-id",
             "price": 3.5,
             "adm": "awesome-markup",
-            "crid": "20",
             "adomain": [
               "awesome.com"
-            ]
+            ],
+            "crid": "20",
+            "w": 1280,
+            "h": 720,
+            "ext": {
+              "prebid": {
+                "type": "video"
+              }
+            }
           },
-          "type": "native"
+          "type": "video"
         }
       ]
     }
