+++ adapters/blasto/blastotest/supplemental/empty-seatbid-array.json
@@ -11,11 +11,17 @@
     "user": {
       "buyeruid": "awesome-user"
     },
-    "site": {
-      "page": "test.com",
+    "app": {
       "publisher": {
         "id": "123456789"
-      }
+      },
+      "cat": [
+        "IAB22-1"
+      ],
+      "bundle": "com.app.awesome",
+      "name": "Awesome App",
+      "domain": "awesomeapp.com",
+      "id": "123456789"
     },
     "imp": [
       {
@@ -60,7 +66,7 @@
             "123.123.123.123"
           ]
         },
-        "uri": "http://host.bizzclick.com/bid?rtb_seat_id=sourceId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "device": {
@@ -72,7 +78,6 @@
           "imp": [
             {
               "id": "some-impression-id",
-              "tagid": "ogTAGID",
               "video": {
                 "mimes": [
                   "video/mp4"
@@ -81,11 +86,18 @@
                 "maxduration": 150,
                 "w": 640,
                 "h": 480
-              }
+              },
+              "tagid": "ogTAGID"
             }
           ],
-          "site": {
-            "page": "test.com",
+          "app": {
+            "id": "123456789",
+            "name": "Awesome App",
+            "bundle": "com.app.awesome",
+            "domain": "awesomeapp.com",
+            "cat": [
+              "IAB22-1"
+            ],
             "publisher": {
               "id": "123456789"
             }
@@ -95,65 +107,28 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
         "status": 200,
         "body": {
           "id": "awesome-resp-id",
-          "seatbid": [
-            {
-              "bid": [
-                {
-                  "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-                  "impid": "some-impression-id",
-                  "price": 3.5,
-                  "adm": "awesome-markup",
-                  "adomain": [
-                    "awesome.com"
-                  ],
-                  "crid": "20",
-                  "w": 1280,
-                  "h": 720,
-                  "ext": {
-                    "prebid": {
-                      "type": "video"
-                    }
-                  }
-                }
-              ],
-              "seat": "bizzclick"
-            }
-          ],
+          "seatbid": [],
           "cur": "USD"
         }
       }
     }
   ],
-  "expectedBidResponses": [
+  "mockResponse": {
+    "status": 200,
+    "body": "invalid response"
+  },
+  "expectedMakeBidsErrors": [
     {
-      "bids": [
-        {
-          "bid": {
-            "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-            "impid": "some-impression-id",
-            "price": 3.5,
-            "adm": "awesome-markup",
-            "adomain": [
-              "awesome.com"
-            ],
-            "crid": "20",
-            "w": 1280,
-            "h": 720,
-            "ext": {
-              "prebid": {
-                "type": "video"
-              }
-            }
-          },
-          "type": "video"
-        }
-      ]
+      "value": "Empty SeatBid array",
+      "comparison": "literal"
     }
   ]
 }
