+++ adapters/blasto/blastotest/supplemental/invalid-blasto-ext-object.json
