+++ adapters/blasto/blastotest/supplemental/invalid-response.json
@@ -27,14 +27,9 @@
       {
         "id": "some-impression-id",
         "tagid": "ogTAGID",
-        "video": {
-          "mimes": [
-            "video/mp4"
-          ],
-          "w": 640,
-          "h": 480,
-          "minduration": 120,
-          "maxduration": 150
+        "banner": {
+          "w": 320,
+          "h": 50
         },
         "ext": {
           "bidder": {
@@ -66,7 +61,7 @@
             "123.123.123.123"
           ]
         },
-        "uri": "http://host.bizzclick.com/bid?rtb_seat_id=sourceId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "device": {
@@ -78,14 +73,9 @@
           "imp": [
             {
               "id": "some-impression-id",
-              "video": {
-                "mimes": [
-                  "video/mp4"
-                ],
-                "minduration": 120,
-                "maxduration": 150,
-                "w": 640,
-                "h": 480
+              "banner": {
+                "w": 320,
+                "h": 50
               },
               "tagid": "ogTAGID"
             }
@@ -107,55 +97,20 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
         "status": 200,
-        "body": {
-          "id": "awesome-resp-id",
-          "seatbid": [
-            {
-              "bid": [
-                {
-                  "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-                  "impid": "some-impression-id",
-                  "price": 3.5,
-                  "adm": "awesome-markup",
-                  "adomain": [
-                    "awesome.com"
-                  ],
-                  "crid": "20",
-                  "w": 1280,
-                  "h": 720
-                }
-              ],
-              "seat": "bizzclick"
-            }
-          ],
-          "cur": "USD"
-        }
+        "body": "invalid response"
       }
     }
   ],
-  "expectedBidResponses": [
+  "expectedMakeBidsErrors": [
     {
-      "bids": [
-        {
-          "bid": {
-            "id": "a3ae1b4e2fc24a4fb45540082e98e161",
-            "impid": "some-impression-id",
-            "price": 3.5,
-            "adm": "awesome-markup",
-            "crid": "20",
-            "adomain": [
-              "awesome.com"
-            ],
-            "w": 1280,
-            "h": 720
-          },
-          "type": "video"
-        }
-      ]
+      "value": "Bad Server Response",
+      "comparison": "literal"
     }
   ]
 }
