+++ adapters/blasto/blastotest/supplemental/status-code-no-content.json
@@ -37,7 +37,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "http://host.bizzclick.com/bid?rtb_seat_id=sourceId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "imp": [
@@ -66,18 +66,15 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
-        "status": 306
+        "status": 204
       }
     }
   ],
   "expectedBidResponses": [],
-  "expectedMakeBidsErrors": [
-    {
-      "value": "Unexpected status code: [ 306 ]. Run with request.debug = 1 for more info",
-      "comparison": "literal"
-    }
-  ]
+  "expectedMakeBidsErrors": []
 }
