+++ adapters/blasto/blastotest/supplemental/status-code-other-error.json
@@ -37,7 +37,7 @@
   "httpCalls": [
     {
       "expectedRequest": {
-        "uri": "http://host.bizzclick.com/bid?rtb_seat_id=sourceId&secret_key=accountId",
+        "uri": "http://host.blasto.ai/bid?rtb_seat_id=sourceId&secret_key=accountId",
         "body": {
           "id": "some-request-id",
           "imp": [
@@ -66,17 +66,19 @@
           },
           "tmax": 1000
         },
-        "impIDs":["some-impression-id"]
+        "impIDs": [
+          "some-impression-id"
+        ]
       },
       "mockResponse": {
-        "status": 503
+        "status": 306
       }
     }
   ],
   "expectedBidResponses": [],
   "expectedMakeBidsErrors": [
     {
-      "value": "Something went wrong, please contact your Account Manager. Status Code: [ 503 ] ",
+      "value": "Unexpected status code: [ 306 ]. Run with request.debug = 1 for more info",
       "comparison": "literal"
     }
   ]
